+++ crates/common_enums/src/enums.rs
@@ -2339,6 +2339,28 @@ pub enum PaymentMethod {
     MobilePayment,
 }
 
+impl PaymentMethod {
+    pub fn is_gift_card(&self) -> bool {
+        match self {
+            Self::GiftCard => true,
+            Self::Card
+            | Self::CardRedirect
+            | Self::PayLater
+            | Self::Wallet
+            | Self::BankRedirect
+            | Self::BankTransfer
+            | Self::Crypto
+            | Self::BankDebit
+            | Self::Reward
+            | Self::RealTimePayment
+            | Self::Upi
+            | Self::Voucher
+            | Self::OpenBanking
+            | Self::MobilePayment => false,
+        }
+    }
+}
+
 /// Indicates the gateway system through which the payment is processed.
 #[derive(
     Clone,
