+++ crates/hyperswitch_domain_models/src/payments/payment_attempt.rs
@@ -633,102 +633,6 @@ impl PaymentAttempt {
         })
     }
 
-    /// Construct the domain model from the ConfirmIntentRequest and PaymentIntent
-    #[cfg(feature = "v2")]
-    pub async fn create_domain_model_for_split_payment(
-        payment_intent: &super::PaymentIntent,
-        cell_id: id_type::CellId,
-        storage_scheme: storage_enums::MerchantStorageScheme,
-        request: &api_models::payments::PaymentsConfirmIntentRequest,
-        encrypted_data: DecryptedPaymentAttempt,
-        split_amount: MinorUnit,
-        attempts_group_id: &id_type::GlobalAttemptGroupId,
-    ) -> CustomResult<Self, errors::api_error_response::ApiErrorResponse> {
-        let id = id_type::GlobalAttemptId::generate(&cell_id);
-        let intent_amount_details = payment_intent.amount_details.clone();
-
-        let attempt_amount_details =
-            intent_amount_details.create_split_attempt_amount_details(request, split_amount);
-
-        let now = common_utils::date_time::now();
-
-        let payment_method_billing_address = encrypted_data
-            .payment_method_billing_address
-            .as_ref()
-            .map(|data| {
-                data.clone()
-                    .deserialize_inner_value(|value| value.parse_value("Address"))
-            })
-            .transpose()
-            .change_context(errors::api_error_response::ApiErrorResponse::InternalServerError)
-            .attach_printable("Unable to decode billing address")?;
-
-        let connector_token = Some(diesel_models::ConnectorTokenDetails {
-            connector_mandate_id: None,
-            connector_token_request_reference_id: Some(common_utils::generate_id_with_len(
-                consts::CONNECTOR_MANDATE_REQUEST_REFERENCE_ID_LENGTH,
-            )),
-        });
-
-        let authentication_type = payment_intent.authentication_type.unwrap_or_default();
-
-        Ok(Self {
-            payment_id: payment_intent.id.clone(),
-            merchant_id: payment_intent.merchant_id.clone(),
-            attempts_group_id: Some(attempts_group_id.to_owned()),
-            amount_details: attempt_amount_details,
-            status: common_enums::AttemptStatus::Started,
-            // This will be decided by the routing algorithm and updated in update trackers
-            // right before calling the connector
-            connector: None,
-            authentication_type,
-            created_at: now,
-            modified_at: now,
-            last_synced: None,
-            cancellation_reason: None,
-            browser_info: request.browser_info.clone(),
-            payment_token: request.payment_token.clone(),
-            connector_metadata: None,
-            payment_experience: None,
-            payment_method_data: None,
-            routing_result: None,
-            preprocessing_step_id: None,
-            multiple_capture_count: None,
-            connector_response_reference_id: None,
-            updated_by: storage_scheme.to_string(),
-            redirection_data: None,
-            encoded_data: None,
-            merchant_connector_id: None,
-            external_three_ds_authentication_attempted: None,
-            authentication_connector: None,
-            authentication_id: None,
-            fingerprint_id: None,
-            charges: None,
-            client_source: None,
-            client_version: None,
-            customer_acceptance: request.customer_acceptance.clone().map(Secret::new),
-            profile_id: payment_intent.profile_id.clone(),
-            organization_id: payment_intent.organization_id.clone(),
-            payment_method_type: request.payment_method_type,
-            payment_method_id: request.payment_method_id.clone(),
-            connector_payment_id: None,
-            payment_method_subtype: request.payment_method_subtype,
-            authentication_applied: None,
-            external_reference_id: None,
-            payment_method_billing_address,
-            error: None,
-            connector_token_details: connector_token,
-            id,
-            card_discovery: None,
-            feature_metadata: None,
-            processor_merchant_id: payment_intent.merchant_id.clone(),
-            created_by: None,
-            connector_request_reference_id: None,
-            network_transaction_id: None,
-            authorized_amount: None,
-        })
-    }
-
     #[cfg(feature = "v2")]
     pub async fn proxy_create_domain_model(
         payment_intent: &super::PaymentIntent,
