+++ crates/router/src/core/split_payments.rs
@@ -5,8 +5,12 @@ use api_models::{
         SplitPaymentMethodDataRequest,
     },
 };
-use common_enums::{CallConnectorAction, SplitTxnsEnabled};
-use common_utils::{ext_traits::OptionExt, id_type, types::MinorUnit};
+use common_enums::CallConnectorAction;
+use common_utils::{
+    ext_traits::{OptionExt, ValueExt},
+    id_type,
+    types::MinorUnit,
+};
 use error_stack::{report, Report, ResultExt};
 use hyperswitch_domain_models::payments::{
     split_payments, HeaderPayload, PaymentConfirmData, PaymentIntent,
@@ -72,11 +76,11 @@ async fn get_payment_method_amount_split(
     payment_intent: &PaymentIntent,
 ) -> RouterResult<split_payments::PaymentMethodAmountSplit> {
     // This function is called inside split payments flow, so its mandatory to have split_payment_method_data
-    let split_payment_method_data = request.split_payment_method_data.clone().ok_or(
-        errors::ApiErrorResponse::MissingRequiredField {
-            field_name: "split_payment_method_data",
-        },
-    )?;
+    let split_payment_method_data = request
+        .split_payment_method_data
+        .clone()
+        .ok_or(errors::ApiErrorResponse::InternalServerError)
+        .attach_printable("split_payment_method_data not found in split payments flow")?;
 
     // The primary/secondary payment method distinction is decided on the backend. Its irrelevant whether a payment_method
     // is received in the top level `payment_method_data` field or inside `split_payment_method_data`
@@ -98,9 +102,9 @@ async fn get_payment_method_amount_split(
         .chain(std::iter::once(payment_method_data))
         .collect();
 
-    let (non_gift_card_pm_data, gift_card_pm_data): (Vec<_>, Vec<_>) = combined_pm_data
+    let (gift_card_pm_data, non_gift_card_pm_data): (Vec<_>, Vec<_>) = combined_pm_data
         .into_iter()
-        .partition(|pm_data| pm_data.payment_method_type != enums::PaymentMethod::GiftCard);
+        .partition(|pm_data| pm_data.payment_method_type.is_gift_card());
 
     // Validate at most one non-gift-card payment method
     if non_gift_card_pm_data.len() > 1 {
@@ -138,14 +142,8 @@ async fn get_payment_method_amount_split(
     )
     .await?;
 
-    let total_balances = balances
-        .iter()
-        .fold(MinorUnit::zero(), |acc, x| acc + x.1.balance);
-
     // TODO: Add surcharge calculation when it is added in v2
     let total_amount = payment_intent.amount_details.calculate_net_amount();
-
-    let remaining_amount = (total_amount - total_balances).max(MinorUnit::zero());
 
     let mut remaining_to_allocate = total_amount;
     let pm_split_amt_tuple: Vec<(PaymentMethodData, MinorUnit)> = gift_card_data_vec
@@ -170,6 +168,7 @@ async fn get_payment_method_amount_split(
             let pm_balance = balances
                 .get(&pm_balance_key)
                 .ok_or(errors::ApiErrorResponse::InternalServerError)
+                .attach_printable("Payment Method Balance not present in Redis")
                 .ok()?;
 
             // Use minimum of available balance and remaining amount
@@ -185,7 +184,7 @@ async fn get_payment_method_amount_split(
 
     // If the gift card balances are not sufficient for payment, use the non-gift card payment method
     // for the remaining amount
-    if remaining_amount > MinorUnit::zero() {
+    if remaining_to_allocate > MinorUnit::zero() {
         let non_gift_card_pm_data = non_gift_card_pm_data
             .ok_or(errors::ApiErrorResponse::InvalidRequestData {
                 message: "Requires additional payment method data".to_string(),
@@ -194,7 +193,7 @@ async fn get_payment_method_amount_split(
 
         Ok(split_payments::PaymentMethodAmountSplit {
             balance_pm_split: pm_split_amt_tuple,
-            non_balance_pm_split: Some((non_gift_card_pm_data, remaining_amount)),
+            non_balance_pm_split: Some((non_gift_card_pm_data, remaining_to_allocate)),
         })
     } else {
         Ok(split_payments::PaymentMethodAmountSplit {
@@ -477,4 +476,103 @@ pub(crate) async fn split_payments_execute_core(
         &profile,
         Some(connector_response_data),
     )
+}
+
+/// Construct the domain model from the ConfirmIntentRequest and PaymentIntent
+#[cfg(feature = "v2")]
+pub async fn create_domain_model_for_split_payment(
+    payment_intent: &PaymentIntent,
+    cell_id: id_type::CellId,
+    storage_scheme: enums::MerchantStorageScheme,
+    request: &api_models::payments::PaymentsConfirmIntentRequest,
+    encrypted_data: hyperswitch_domain_models::payments::payment_attempt::DecryptedPaymentAttempt,
+    split_amount: MinorUnit,
+    attempts_group_id: &id_type::GlobalAttemptGroupId,
+) -> common_utils::errors::CustomResult<domain::PaymentAttempt, errors::ApiErrorResponse> {
+    let id = id_type::GlobalAttemptId::generate(&cell_id);
+    let intent_amount_details = payment_intent.amount_details.clone();
+
+    let attempt_amount_details =
+        intent_amount_details.create_split_attempt_amount_details(request, split_amount);
+
+    let now = common_utils::date_time::now();
+
+    let payment_method_billing_address = encrypted_data
+        .payment_method_billing_address
+        .as_ref()
+        .map(|data| {
+            data.clone()
+                .deserialize_inner_value(|value| value.parse_value("Address"))
+        })
+        .transpose()
+        .change_context(errors::ApiErrorResponse::InternalServerError)
+        .attach_printable("Unable to decode billing address")?;
+
+    let connector_token = Some(diesel_models::ConnectorTokenDetails {
+        connector_mandate_id: None,
+        connector_token_request_reference_id: Some(common_utils::generate_id_with_len(
+            hyperswitch_domain_models::consts::CONNECTOR_MANDATE_REQUEST_REFERENCE_ID_LENGTH,
+        )),
+    });
+
+    let authentication_type = payment_intent.authentication_type.unwrap_or_default();
+
+    Ok(domain::PaymentAttempt {
+        payment_id: payment_intent.id.clone(),
+        merchant_id: payment_intent.merchant_id.clone(),
+        attempts_group_id: Some(attempts_group_id.to_owned()),
+        amount_details: attempt_amount_details,
+        status: common_enums::AttemptStatus::Started,
+        // This will be decided by the routing algorithm and updated in update trackers
+        // right before calling the connector
+        connector: None,
+        authentication_type,
+        created_at: now,
+        modified_at: now,
+        last_synced: None,
+        cancellation_reason: None,
+        browser_info: request.browser_info.clone(),
+        payment_token: request.payment_token.clone(),
+        connector_metadata: None,
+        payment_experience: None,
+        payment_method_data: None,
+        routing_result: None,
+        preprocessing_step_id: None,
+        multiple_capture_count: None,
+        connector_response_reference_id: None,
+        updated_by: storage_scheme.to_string(),
+        redirection_data: None,
+        encoded_data: None,
+        merchant_connector_id: None,
+        external_three_ds_authentication_attempted: None,
+        authentication_connector: None,
+        authentication_id: None,
+        fingerprint_id: None,
+        charges: None,
+        client_source: None,
+        client_version: None,
+        customer_acceptance: request
+            .customer_acceptance
+            .clone()
+            .map(masking::Secret::new),
+        profile_id: payment_intent.profile_id.clone(),
+        organization_id: payment_intent.organization_id.clone(),
+        payment_method_type: request.payment_method_type,
+        payment_method_id: request.payment_method_id.clone(),
+        connector_payment_id: None,
+        payment_method_subtype: request.payment_method_subtype,
+        authentication_applied: None,
+        external_reference_id: None,
+        payment_method_billing_address,
+        error: None,
+        connector_token_details: connector_token,
+        id,
+        card_discovery: None,
+        feature_metadata: None,
+        processor_merchant_id: payment_intent.merchant_id.clone(),
+        created_by: None,
+        connector_request_reference_id: None,
+        network_transaction_id: None,
+        authorized_amount: None,
+    })
 }
